feat: added compatibility with mongodb
